claire_agent/requirements.txt
# requirements.txt
streamlit
python-dotenv
langchain
langchain-community
pydantic
chromadb
sentence-transformers # HuggingFaceEmbeddings에 필요